[rename]: remote의 'service' package -> 'api'
